Enhanced root finding and matrix operations
As a few more tests were conducted with
root finding (with Newton-Raphson's and
Modified Newton-Raphson's), multiple
roots can be determined, given that
enough rounds and iterations are provided
for the algorithm.

Along with root finding, more operations
were added to the matrix class, allowing
indexed construction, as well as the
addition, multiplication and subtraction
operators.