Restored Node Functionality With Stree
The convertion process from stree to node
has been filled out (for now) and is at
the moment functional; the root finding
test works as before to prove this.

The convertion process is still rough
and in work; with changes forces by the
config class and future merging of
variable and function stack, the
convertion process should be both
more fluid, simple and less error-prone
to changes the user would like to implement.