Fixed Bug With Operations
A bug that appeared whenever
a functor or tree object (node)
was created globally was
fixed in this patch.

The fix required the reworking
(redesigning) of the operation
class, and the defaults class
needed to be scrapped; the idea
of using reference to static
instances would simply not
hold out, especially not towards
the future. New operations needed
to be allocated instead of referenced.

Additionally, to compensate for
the loss of the defaults class,
a new class, the config class,
was added in this patch. Similar
to the defaults class, the client
(user) could create and specify
their own set of operations, as
long as the parser is able to read
them. The ability to customize real
space constants such as one, zero and
the negative are also a part of
this class.