Elements and utilities
A subclass of matrix, the element class, was created in this patch with
the intention of representing vectors and individual elements of a given
vector space.

Additionaly, as utilities module was added (as a namespace, not a
class), which is intended to be a collection of common/practical methods
in math/computer science, such as the Gram Scmidt process.