Table fluidity
This patch is made mainly to add new constructors to thte table class
which allow for users to add arbitrary symbols (variables and functors)
to the table at the time of construction.

Other additions were made in this patch such as differentiation of
exponential functions. Differentiation (and compression) by division is
still under work.