Completed UI
The design, framework and backbone of the webpage is complete, with
features of graphing (including implicit graphing of single-variable
functions).

As an added bonus for this patch, the functor class was finally cleaned
up.

Next up on the list of tasks is the chain rule, as well as the
completion of differention (all rules).