More accurate Bernoulli numbers
With the creation of the rational class comes the ability to represent
rationla numbers without losing precision. This advantage over the
standard double or floating point data point is utilized in the
alternate form of the Bernoulli sequence generator.

Errors and unwanted results regarding the GCD/LCM functions where also
fixed in this patch.