Complex numbers and linear differential equations
Complex numbers were added as an additional number set. This class is a
WRAPPER class of the std::complex class, with additional methods that
are intended to make it more convenient to use when compared to the
standard complex class.

Solving for the (complex) roots of a polynomial was also added into the
polynomial class, using the Durand Kerner method. With this the
solutions of homogenous linear differential equations could and can be
found.