Switch to double type
A large change was made which mainly consisted of switching from the
singular datatype-based operations to a duplex datatype based operation
scheme. Now, instead of a accepting a single parameter, the functor and
node class (along with other affected class) require an additional
template parameter for specifying the integer set.