Result of training on MNIST hand written digits set
In 17 epochs, the neural network was able to achieve an accuracy of 86%
on the the training data. From the 18th to 20th epoch, the accuracy was
hanging around this value, due to the fact that the learning rate was
increasing instead of decreasing.

The neural network was using momentum gradient descent and a tweaked
performance learning rate scheduler, which is probably the cause of the
dying improvement.