Added the ability to multithread neural net training
Additionally ran two 100 epoch training sessions with two slightly
altered neural nets. Both ran with the Nesterov optimizer and an
exponential learning rate scheduler, with the only difference being the
hyperparameters involved with the scheduler. The two nets achieved
accuracies of 91.68% and 92.63% on the training set, with the larger
accuracy going to the net with a schedular which converged faster.