Established basic methods for stringing a network
The new NetNode class represents a 'layer' in an ANN. It can take in
multiple inputs and outputs. Using the overloaded '<<' operator, these
inputs and outputs can be piped. For example, if n1 and n2 are two
NetNodes, then the statement n1[1] << n2[0] pipes the first output of n2
to the second input of n1 (using 0-indexing). The pipes are created
dynamically, so the user does not need to worry about allocating space
for them.

An ANN will be constructed from a set of input NetNodes, which require
no inputs (the ANN class is not yet implemented, this is a sketch of its
implementation). Users can call the appropriate method for evaluating
the network with a set of inputs, but the outputs must be manually
retrieved, meaning that nothing is returned from the evalution. Most
likely these outputs will be retrieved with the index operator (ann[0],
etc.).

Also added DDQN learning to the RL simulation. The RL simulation
directory will be extrapolated into a new repositoty as it is becoming a
standalone project.