Reworking attirbutes to use MethodTables
Each Token object will hold a pointer to a MethodTable (similar to
vtables in C++). This saves a lot of storage: compared to before, where
each object would hold the table information (and documentation), we
only have one pointer. Canonical use of the MethodTable is as a static
variable of the subclass, whos address is passed to the Token
constructor.