Versioning and bug fixes
The interpreter and any libraries it compiles will be linted with a
version title from now on. This version title contains the version
number (ie. v1.0.0) and the date of its compilation (which is in a bit
of a buggy state).

Fixed indexing to account for both references and values. The statement
`p[0] = 6` will assign through the iterator and `x = p[0]` will retrieve
through the value.